CR-3 Push client side refacto I'm not happy with the "syncOnServerCall in the init : nothing done. Check SRP then refacto : WiP Details for using all browser APIs which allow us to use Push Notification are way numberous, so I'm trying to delegate to a PushService object.
Next step : continue checking SRP then refacto
